Move ahrs stuff out of genIfaces.ahrs and into ahrs
