+++ src/main/java/frc/team449/drive/unidirectional/commands/AHRS/NavXDriveStraight.java
@@ -8,7 +8,7 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.Subsystem;
 import frc.team449.ahrs.PIDAngleController;
 import frc.team449.drive.unidirectional.DriveUnidirectional;
-import frc.team449.generalInterfaces.ahrs.SubsystemAHRS;
+import frc.team449.ahrs.SubsystemAHRS;
 import frc.team449.oi.unidirectional.tank.OITank;
 import org.jetbrains.annotations.NotNull;
