+++ src/main/java/frc/team449/drive/unidirectional/commands/AHRS/NavXTurnToAngleRelative.java
@@ -5,7 +5,7 @@ import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj2.command.Subsystem;
 import frc.team449.ahrs.PIDAngleController;
 import frc.team449.drive.unidirectional.DriveUnidirectional;
-import frc.team449.generalInterfaces.ahrs.SubsystemAHRS;
+import frc.team449.ahrs.SubsystemAHRS;
 import frc.team449.other.Clock;
 import frc.team449.other.Util;
 import org.jetbrains.annotations.NotNull;
